Refactor repository tests and update entity creation validation
